+++ src/main/java/net/spell_engine/client/gui/Drawable.java
@@ -5,7 +5,7 @@ import net.minecraft.util.Identifier;
 
 public class Drawable {
     public enum Anchor { LEADING, TRAILING, CENTER }
-    public record Component(Draw draw, Texture texture) {
+    public record Component(DrawRect draw, Texture texture) {
         public void draw(DrawContext context, int x, int y, Anchor hAnchor, Anchor vAnchor) {
             switch (hAnchor) {
                 case LEADING -> {
@@ -47,6 +47,6 @@ public class Drawable {
             context.drawTexture(texture().id, x, y, draw().u, draw().v, width, draw().height, texture().width, texture().height);
         }
     }
-    public record Draw(int u, int v, int width, int height) {}
+    public record DrawRect(int u, int v, int width, int height) {}
     public record Texture(Identifier id, int width, int height) {}
 }